subindo todas as rotas com autentificação

+++ src/controller/api_budgets.js
@@ -3,7 +3,8 @@ const { Budgets } = require('../database/models');
 const api_Budgets = {
     get: async( req, res) => {
         try {
-            const response = await Budgets.findAll()
+            const user_id = req.tokenDecoded.id
+            const response = await Budgets.findAll({where:{user_id:user_id}})
 
             return res.status(200).json(response);
         } catch (error) {
@@ -13,7 +14,9 @@ const api_Budgets = {
     },
     post: async( req, res) => {
         try {
+            req.body.user_id = req.tokenDecoded.id
             const response = await Budgets.create(req.body)
+
             return res.status(201).json({response});
         } catch (error) {
             console.log(error);
@@ -23,8 +26,9 @@ const api_Budgets = {
     put: async( req, res) => {
         try {
             const {id} = req.body
+            const user_id = req.tokenDecoded.id
 
-            let response = await Budgets.findByPk(id)
+            let response = await Budgets.findOne({where:{id:id, user_id:user_id}})
             delete req.body.id
             let updated = await response.update(req.body)
             
@@ -37,8 +41,9 @@ const api_Budgets = {
     delete: async( req, res) => {
         try {
             const {id} = req.body
+            const user_id = req.tokenDecoded.id
 
-            let response = await Budgets.destroy({where:{id:id}})
+            let response = await Budgets.destroy({where:{id:id, user_id:user_id }})
             
             return res.status(200).json(response);
         } catch (error) {
@@ -49,8 +54,9 @@ const api_Budgets = {
 
     id_budgets:async (req, res) => {
         try {
+            const user_id = req.tokenDecoded.id
             let response = await Budgets.findOne({
-                where:{id: Number(req.body.id) },
+                where:{id: Number(req.body.id), user_id:user_id },
                 include:'category'
             });
 
@@ -60,22 +66,6 @@ const api_Budgets = {
             return res.status(500).json({msg:"Original Error [id_budgets.category-findOne]GET status-500 client-server error!"});
         }
     },
-    
-    user_budgets:async (req, res) => {
-        try {
-            let response = await Budgets.findAll({
-                where:{user_id: Number(req.body.id) },
-                include:'category'
-            });
-
-            return res.status(200).json(response);
-        } catch (error) {
-            console.log(error);
-            return res.status(500).json({msg:"Original Error [user_budgets.category-findAll]GET status-500 client-server error!"});
-        }
-    },
-    
-
 }
 
 module.exports = api_Budgets;
