+++ src/controller/api_category.js
@@ -3,7 +3,8 @@ const { Category } = require('../database/models');
 const api_category = {
     get: async( req, res) => {
         try {
-            const response = await Category.findAll()
+            const user_id = req.tokenDecoded.id
+            const response = await Category.findAll({where:{user_id:user_id}})
 
             return res.status(200).json(response);
         } catch (error) {
@@ -13,7 +14,9 @@ const api_category = {
     },
     post: async( req, res) => {
         try {
+            req.body.user_id = req.tokenDecoded.id
             const response = await Category.create(req.body)
+            
             return res.status(201).json({response});
         } catch (error) {
             console.log(error);
@@ -23,8 +26,9 @@ const api_category = {
     put: async( req, res) => {
         try {
             const {id} = req.body
+            const user_id = req.tokenDecoded.id
 
-            let response = await Category.findByPk(id)
+            let response = await Category.findOne({where:{id:id, user_id:user_id}})
             delete req.body.id
             let updated = await response.update(req.body)
             
@@ -37,8 +41,8 @@ const api_category = {
     delete: async( req, res) => {
         try {
             const {id} = req.body
-
-            let response = await Category.destroy({where:{id:id}})
+            const user_id = req.tokenDecoded.id
+            let response = await Category.destroy({where:{id:id, user_id:user_id}})
             
             return res.status(200).json(response);
         } catch (error) {
@@ -50,25 +54,13 @@ const api_category = {
     /* advance search */
     id_category:async (req, res) => {
         try {
-            let response = await Category.findOne({ where: { id: Number( req.body.id ) } });
+            const user_id = req.tokenDecoded.id
+            let response = await Category.findOne({ where: { id: Number( req.body.id ), user_id:user_id } });
 
             return res.status(200).json(response);
         } catch (error) {
             console.log(error);
             return res.status(500).json({msg:"Original Error [id_category-findOne]GET status-500 client-server error!"});
-        }
-    },
-    
-    user_category:async (req, res) => {
-        try {
-            let response = await Category.findAll({
-                where:{user_id:Number(req.body.id)}
-            });
-
-            return res.status(200).json(response);
-        } catch (error) {
-            console.log(error);
-            return res.status(500).json({msg:"Original Error [user_category-findAll]GET status-500 client-server error!"});
         }
     },
 }
