+++ src/controller/api_settings.js
@@ -3,7 +3,8 @@ const { Settings } = require('../database/models');
 const api_Settings = {
     get: async( req, res) => {
         try {
-            const response = await Settings.findAll()
+            const user_id = req.tokenDecoded.id
+            const response = await Settings.findAll({where:{user_id:user_id}})
 
             return res.status(200).json(response);
         } catch (error) {
@@ -13,6 +14,7 @@ const api_Settings = {
     },
     post: async( req, res) => {
         try {
+            req.body.user_id = req.tokenDecoded.id
             const response = await Settings.create(req.body)
             return res.status(201).json({response});
         } catch (error) {
@@ -37,8 +39,9 @@ const api_Settings = {
     delete: async( req, res) => {
         try {
             const {id} = req.body
+            const user_id = req.tokenDecoded.id
 
-            let response = await Settings.destroy({where:{id:id}})
+            let response = await Settings.destroy({where:{id:id, user_id:user_id}})
             
             return res.status(200).json(response);
         } catch (error) {
@@ -50,25 +53,13 @@ const api_Settings = {
     /* advance search */
     id_settings:async (req, res) => {
         try {
-            let response = await Settings.findOne({ where: { id: Number( req.body.id ) } });
+            const user_id = req.tokenDecoded.id
+            let response = await Settings.findOne({ where: { id: Number( req.body.id ), user_id:user_id } });
 
             return res.status(200).json(response);
         } catch (error) {
             console.log(error);
             return res.status(500).json({msg:"Original Error [id_settings-findOne]GET status-500 client-server error!"});
-        }
-    },
-    
-    user_settings:async (req, res) => {
-        try {
-            let response = await Settings.findAll({
-                where:{user_id:Number(req.body.id)}
-            });
-
-            return res.status(200).json(response);
-        } catch (error) {
-            console.log(error);
-            return res.status(500).json({msg:"Original Error [user_settings-findAll]GET status-500 client-server error!"});
         }
     },
 }
