+++ src/controller/api_support.js
@@ -3,7 +3,8 @@ const { Support_tickets } = require('../database/models');
 const api_Support_tickets = {
     get: async( req, res) => {
         try {
-            const response = await Support_tickets.findAll()
+            const user_id = req.tokenDecoded.id
+            const response = await Support_tickets.findAll({where:{user_id:user_id}})
 
             return res.status(200).json(response);
         } catch (error) {
@@ -13,7 +14,10 @@ const api_Support_tickets = {
     },
     post: async( req, res) => {
         try {
+            req.body.user_id = req.tokenDecoded.id
+
             const response = await Support_tickets.create(req.body)
+
             return res.status(201).json({response});
         } catch (error) {
             console.log(error);
@@ -23,9 +27,11 @@ const api_Support_tickets = {
     put: async( req, res) => {
         try {
             const {id} = req.body
+            const user_id = req.tokenDecoded.id
 
-            let response = await Support_tickets.findByPk(id)
+            let response = await Support_tickets.findOne({where:{id:id, user_id: user_id}})
             delete req.body.id
+
             let updated = await response.update(req.body)
             
             return res.status(200).json(updated);
@@ -37,8 +43,9 @@ const api_Support_tickets = {
     delete: async( req, res) => {
         try {
             const {id} = req.body
+            const user_id = req.tokenDecoded.id
 
-            let response = await Support_tickets.destroy({where:{id:id}})
+            let response = await Support_tickets.destroy({where:{id:id, user_id:user_id}})
             
             return res.status(200).json(response);
         } catch (error) {
@@ -50,25 +57,13 @@ const api_Support_tickets = {
     /* advance search */
     id_support:async (req, res) => {
         try {
-            let response = await Support_tickets.findOne({ where: { id: Number( req.body.id ) } });
+            const user_id = req.tokenDecoded.id
+            let response = await Support_tickets.findOne({ where: { id: Number( req.body.id ), user_id:user_id } });
 
             return res.status(200).json(response);
         } catch (error) {
             console.log(error);
             return res.status(500).json({msg:"Original Error [id_support-findOne]GET status-500 client-server error!"});
-        }
-    },
-    
-    user_support:async (req, res) => {
-        try {
-            let response = await Support_tickets.findAll({
-                where:{user_id:Number(req.body.id)}
-            });
-
-            return res.status(200).json(response);
-        } catch (error) {
-            console.log(error);
-            return res.status(500).json({msg:"Original Error [user_support-findAll]GET status-500 client-server error!"});
         }
     },
 }
